% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_github.R
\name{check_github}
\alias{check_github}
\title{check_github}
\usage{
check_github(pkg)
}
\arguments{
\item{pkg}{}
}
\value{
list with the comparison between what is in the system and
 what is in github
}
\description{
check latest github version of R package
}
\examples{
# No example - cf - 
## https://github.com/r-lib/pkgcache#using-pkgcache-in-cran-packages

# check_github('vidonne/unhcrthemes')
}
