% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcrverse.R
\name{unhcrverse}
\alias{unhcrverse}
\title{unhcrverse}
\usage{
unhcrverse(dev = FALSE)
}
\arguments{
\item{dev}{Boolean TRUE / FALSE to indicate if the dev version should be used}
}
\value{
nothing only messages

list
}
\description{
check and load latest version of unhcrverse
}
\examples{

# install.packages("pak")
# pak::pkg_install("unhcRverse/unhcRverse")

# unhcrverse::unhcrverse() 

## If you want to contribute to package development, 
# which you are probably if you are currently checking this test :)
# use the version from the developers
# unhcrverse::unhcrverse(dev = TRUE)
}
