% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_github.R
\name{update_github}
\alias{update_github}
\title{update_github}
\usage{
update_github(pkg)
}
\arguments{
\item{pkg}{}
}
\value{
nothing just messaging what is happening
}
\description{
Update package from github
}
\examples{
# No example - cf - 
## https://github.com/r-lib/pkgcache#using-pkgcache-in-cran-packages

# update_github(pkg = 'vidonne/unhcrthemes')


     
}
